[icxtunnel] implement autodiscovery and include a hidden standalone dev relay command for testing (#67)
Now connects to a seed relay and bootstraps the list of relays from that
connection.

Also can finally spin up a standalone dev relay,

```shell
sudo -E ./apoxy alpha tunnel relay
```

And connect to it with,

```shell
./apoxy alpha tunnel run --agent demo --name dev --relay-addr=localhost:6081 --token=letmein --insecure-skip-verify
```

This uncovered a bunch of overlooked bugs that have also been fixed in
this PR.

I think it might make sense in the future to include an alternative icx
server router implementation that uses the cross platform tun device
library and the unprivileged bifurcated/l2pc logic that we use in the
agent code. This should allow running a standalone dev relay on mac osx
etc.

Super chuffed to see this finally coming together!

Still gotta implement the kubernetes autodiscovery logic for the lambda
use cases etc, that'll be coming soon (and is super straight forward).