[icxtunnel] update icx for allowed routes semantics (#80)
So that redundant routes will work as expected. Not to mention that I
was never a big fan of `allowed_ips` term and found it rather ambiguous
(eg. what ip?).

Also for some reason stats have to be signed int64 or k8s craps itself.

Oh and I forgot to enable keepalives in the routers 🤦‍♂️

Still easily pulls 1gbit/s+ from the socks/usermode client.